+++ example/distinct_value_connectable_stream_example.dart
@@ -8,7 +8,7 @@ class CounterBloc {
   final void Function(int) decrement;
 
   /// Outputs
-  final ValueStream<int> state;
+  final DistinctValueStream<int> state$;
 
   /// Clean up
   final void Function() dispose;
@@ -16,7 +16,7 @@ class CounterBloc {
   CounterBloc._({
     @required this.increment,
     @required this.decrement,
-    @required this.state,
+    @required this.state$,
     @required this.dispose,
   });
 
@@ -30,18 +30,18 @@ class CounterBloc {
     ];
     final state$ = Rx.merge(streams)
         .scan<int>((acc, e, _) => acc + e, 0)
-        .publishValueSeededDistinct(seedValue: 0);
+        .publishValueDistinct(0);
 
     final subscription = state$.connect();
 
     return CounterBloc._(
       increment: incrementController.add,
       decrement: decrementController.add,
-      state: state$,
+      state$: state$,
       dispose: () async {
-        await subscription.cancel();
-        await Future.wait(
+        await Future.wait<void>(
             [incrementController, decrementController].map((c) => c.close()));
+        await subscription.cancel();
       },
     );
   }
@@ -50,7 +50,9 @@ class CounterBloc {
 void main() async {
   final counterBloc = CounterBloc();
 
-  final listen = counterBloc.state.listen((i) => print('[LOGGER] state=$i'));
+  print('[LOGGER] state=${counterBloc.state$.value}');
+  final listen = counterBloc.state$.listen((i) => print('[LOGGER] state=$i'));
+
   counterBloc
     ..increment(0)
     ..increment(2)
@@ -65,8 +67,8 @@ void main() async {
     ..increment(0)
     ..increment(0);
 
-  await Future.delayed(Duration(seconds: 1));
-  print(counterBloc.state.value);
+  await Future<void>.delayed(Duration(seconds: 1));
+  print(counterBloc.state$.value);
 
   await listen.cancel();
   await counterBloc.dispose();