+++ lib/distinct_value_connectable_stream.dart
@@ -1,4 +1,6 @@
 library distinct_value_connectable_stream;
 
 export 'src/distinct_value_connectable_stream.dart';
+export 'src/distinct_value_stream.dart';
 export 'src/extension_methods.dart';
+export 'src/value_subject.dart';