+++ lib/src/distinct_value_stream.dart
@@ -1,0 +1,13 @@
+import 'package:rxdart/streams.dart';
+
+/// An [Stream] that provides synchronous access to the last emitted item,
+/// and two consecutive values are not equal.
+/// The equality between previous data event and current data event is determined by [equals].
+abstract class DistinctValueStream<T> extends ValueStream<T> {
+  /// Determined equality between previous data event and current data event.
+  bool Function(T, T) get equals;
+
+  /// Default [equals] function.
+  /// Use '==' operator on the last provided data element.
+  static final defaultEquals = (Object lhs, Object rhs) => lhs == rhs;
+}